Control Panel Buttons (crappy GUI ) and login screen
Control Panel which contains a buttons for each sections of the program.
Need to design the look and feel of the Control Panel.
Need to clean the code up to match and possible move to another class

Login window read for a backend connection.
Needing hashing of password, comparing password, setting users access level.