Salt + Hash Server Side Implemented + Default User Added
Passwords in the database are now stored as a salt + hash of the password for that user. This is then used to compare with the entered password at login.

A default user has been added to the queries file if db isn't created.

A few minor coding clean ups and typos corrected