User Authentication Token Implemented
When the user logs in successfully, the server will generate a token and pass the information into a hashmap which stores the key as the username of the user and the value as a string array that contains the auth token and the expiry time. After this has been done, the server will then return the auth token to the client. For future code implementation of actions for the user, a function has been written that authenticates the user and checks for the following things in order: 1. if the session for the user exists 2. check that the token passed from the client matches up with the token stored in the server to ensure there was no tempering 3. checks the token hasn't expired yet