Create New User Implemented
Administrator can now create a user and select their permissions. The administrator cannot create a user that already exists or a user with missing fields. If they attempt to, an error dialog will appear stating they must fill all fields or choose a different username.

The current list of a users is requested to the server once a new user has been requested to be created to check the current username isn't taken.

This was implemented with unit tests to ensure the logic flowed correctly