User Feedback On Actions Implemented
User interaction with the Control Panel now involves feedback for the user based on whether their interaction was successful or not. This is based on whether the user has a valid auth token for their username. If the feedback is an error, the user is logged out due to the failed auth check